# docs/api.yaml
openapi: 3.0.0
info:
  title: "Data extractor api"
  version: "1.0.0"
  contact:
    name: Orlando Burli
    email: orlando.burli@gmail.com
    url: https://orlandoburli.com.br/
tags:
  - name: jobs
    description: Jobs api
paths:
  /jobs:
    post:
      operationId: createJob
      summary: Starts a job scrappy to collect data
      tags: [jobs]
      requestBody:
        description: Job request data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobRequest'
      responses:
        '201':
          description: Job successfully created
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/JobCreationResponse'
        '400':
          description: Job creation failed - bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobErrorResponse'
  /jobs/{id}:
    get:
      operationId: getJobStatus
      summary: Return the status of the job
      tags: [jobs]
      responses:
        '200':
          description: Data found
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/JobGetResponse'
        '404':
          description: Job not found

components:
  schemas:
    JobRequest:
      type: object
      properties:
        yearStart:
          type: integer
          format: int64
          example: 1970
          description: First year to start the data collector
        yearEnd:
          type: integer
          format: int64
          example: 2022
          description: Last year to start the data collector

    JobCreationResponse:
      type: object
      properties:
        status:
          type: string
          description: Result of jobs creation
        results:
          type: array
          description: List of results for multiple jobs created
          items:
            type: object
            properties:
              dag_run_id:
                type: string
                description: Job ID
                example: 123456
              status:
                type: string
                description: Status of the job/dag run
              params:
                type: object
                description: Params used to create the job
                properties:
                  year:
                    type: integer
                    description: Year of data extraction
                  option:
                    type: string
                    description: Option tag data used to start extraction
    JobErrorResponse:
      type: object
      properties:
        detail:
          type: array
          description: Details of the failure
          items:
            type: object
            properties:
              type:
                type: string
                description: Type of failure
              log:
                type: string
                description: Location of the failure
              msg:
                type: string
                description: Error message
              input:
                type: string
                description: Input that triggers the error
    JobGetResponse:
      type: object
      properties:
        status: 
          type: string
          description: Status of the job
          example: In progress